Support flexible CKKS scale and modulus chains and redesign benchmarks
Decouple CKKS scales from prime widths, support multi-prime depth
groups, and rescale by complete group products.

Improve Bootstrap scale composition, material reuse, and intermediate
lifetimes. Update documentation and examples to the current APIs.

Redesign benchmarks around cross-hardware workload comparisons,
with methodology, execution guidance, and result submissions.

Bump version to 0.25.0.

+++ fhelium/backend/ckks/codec/_embedding.py
@@ -234,7 +234,7 @@ def encode_slots(
         device=device,
         generator=generator,
     )
-    return rng.randround(coefficients * float(scale))
+    return rng.randround(coefficients * float(scale), dtype=torch.int64)
 
 
 def decode_slots(
